+++ src/bl64-arc.bash
@@ -4,7 +4,7 @@
 # Author: serdigital64 (https://github.com/serdigital64)
 # License: GPL-3.0-or-later (https://www.gnu.org/licenses/gpl-3.0.txt)
 # Repository: https://github.com/serdigital64/bashlib64
-# Version: 1.4.1
+# Version: 1.5.0
 #######################################
 
 #######################################
@@ -21,32 +21,25 @@
 #   STDOUT: None
 #   STDERR: tar or lib error messages
 # Returns:
-#   BL64_ARC_ERROR_MISSING_PARAMETER
 #   BL64_ARC_ERROR_INVALID_DESTINATION
 #   tar error status
 #######################################
 function bl64_arc_open_tar() {
+  bl64_dbg_lib_show_function "$@"
   local source="$1"
   local destination="$2"
   local -i status=0
 
-  if [[ -z "$source" || -z "$destination" ]]; then
-    bl64_msg_show_error "$_BL64_ARC_TXT_MISSING_PARAMETER (source,destination)"
-    # shellcheck disable=SC2086
-    return $BL64_ARC_ERROR_MISSING_PARAMETER
-  fi
-
-  if [[ ! -d "$destination" ]]; then
-    bl64_msg_show_error "$_BL64_ARC_TXT_DST_NOT_DIRECTORY ($destination)"
-    # shellcheck disable=SC2086
-    return $BL64_ARC_ERROR_INVALID_DESTINATION
-  fi
+  bl64_check_parameter 'source' &&
+    bl64_check_parameter 'destination' &&
+    bl64_check_directory "$destination" ||
+    return $?
 
   cd "$destination" || return 1
 
   case "$BL64_OS_DISTRO" in
   ${BL64_OS_UB}-* | ${BL64_OS_DEB}-* | ${BL64_OS_FD}-* | ${BL64_OS_CNT}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_OL}-*)
-    "$BL64_OS_CMD_TAR" \
+    "$BL64_ARC_CMD_TAR" \
       --overwrite \
       --extract \
       --no-same-owner \
@@ -59,7 +52,7 @@ function bl64_arc_open_tar() {
     status=$?
     ;;
   ${BL64_OS_ALP}-*)
-    "$BL64_OS_CMD_TAR" \
+    "$BL64_ARC_CMD_TAR" \
       x \
       --overwrite \
       -f "$source" \
@@ -68,7 +61,7 @@ function bl64_arc_open_tar() {
     status=$?
     ;;
   ${BL64_OS_MCOS}-*)
-    "$BL64_OS_CMD_TAR" \
+    "$BL64_ARC_CMD_TAR" \
       --extract \
       --no-same-owner \
       --preserve-permissions \
@@ -78,10 +71,14 @@ function bl64_arc_open_tar() {
       --file="$source"
     status=$?
     ;;
-  *) bl64_msg_show_unsupported ;;
+  *)
+    bl64_check_show_unsupported
+    # shellcheck disable=SC2086
+    return $BL64_LIB_ERROR_APP_INCOMPATIBLE
+    ;;
   esac
 
-  ((status == 0)) && $BL64_OS_ALIAS_RM_FILE "$source"
+  ((status == 0)) && bl64_fs_rm_file "$source"
 
   return $status
 }